Refactor: renamed marker package to markers
