+++ packages/database/prefix.go
@@ -7,8 +7,8 @@ const (
 	PrefixHealth
 	// PrefixMessageLayer defines the storage prefix for the message layer.
 	PrefixMessageLayer
-	// PrefixMarker defines the storage prefix for the markers used to optimize structural checks in the tangle.
-	PrefixMarker
+	// PrefixMarkers defines the storage prefix for the markers used to optimize structural checks in the tangle.
+	PrefixMarkers
 	// PrefixValueTransfers defines the storage prefix for value transfer.
 	PrefixValueTransfers
 )
